refactor: rework build website
