--- GCanvas/docs/About.md
@@ -1,46 +1,0 @@
-GCanvas start since 2013 in alibaba. At that time, canvas animation running on a mobile app was really a nightmare(at least in most cases). A few severe problems prevented developers utilizing the canvas in a wide way:
-- slow performance
-- bad compatibility(especially on Android)
-- lack of completeness of canvas APIs
-
-To solve the problems mentioned above, we decided to write a cross-platform solutions and provide APIs in JavaScript fully aligned with HTML5 canvas APIs. To ensure the performance, C++ with OpenGL ES appears to be the only way.
-
-Time has passed, powerful mobile hybrid frameworks such as ReactNative/Weex have become popular, but Canvas/WebGL has been scarcely implemented on these frameworks. So we have GCanvas integrated into the most popular framework to provide high performance canvas APIs.
-
-Contributors:
-腾渊 兵长 芒格 许凡 米悠 济宇 星宝 宬豪 济翼 怀听 凝砺 荣木 清滢 韩锷 鸾萱 二同 韦青 胡帅 叶斋 寒泉 [<I want my name appears here>]()
-
-Great appreciate all GCanvas contributors.
-
-Active maintainers:
-<table>
-  <tbody>
-    <tr>
-      <td align="center" valign="top">
-        <img width="128" height="128" src="https://avatars3.githubusercontent.com/u/1937795?s=400&u=7814600b27a1270921892f44588f22754fa297d0&v=4">
-        <br>
-        <a href="https://github.com/fREEDtICE">@ertong</a>
-      </td>
-      <td align="center" valign="top">
-        <img width="128" height="128" src="https://avatars0.githubusercontent.com/u/1164433?s=460&v=4">
-        <br>
-        <a href="https://github.com/jwxbond">@jwxbond</a>
-      </td>
-      <td align="center" valign="top">
-        <img width="128" height="128" src="https://avatars0.githubusercontent.com/u/10558069?s=460&v=4">
-        <br>
-        <a href="https://github.com/sylonhu">@sylonhu</a>
-      </td>
-      <td align="center" valign="top">
-        <img width="128" height="128" src="https://avatars1.githubusercontent.com/u/2938447?s=400&v=4">
-        <br>
-        <a href="https://github.com/xieguanglei">@xieguanglei</a>
-      </td>
-      <td align="center" valign="top">
-        <img width="128" height="128" src="https://avatars2.githubusercontent.com/u/726566?s=400&v=4">
-        <br>
-        <a href="https://github.com/wintercn">@wintercn</a>
-      </td>
-     </tr>
-  </tbody>
-</table>
