--- GCanvas/docs/Reference_Graphics_2D.md
@@ -1,88 +1,0 @@
-### Usage
-
-For gcanvas is designed to follow web canvas API, find details of each API in
-
-[https://developer.mozilla.org/en-US/docs/Web/API/Canvas\_API](https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API)
-###
-
-### Support
-
-Legend
-
-* Implemented: API is implemented .
-* **Undefined**: API is not defined.
-* _Empty_: API is defined but not implemented.
-
-
-#### 1.Context
-| API Name                 | API Type                | Status        |
-| ------------------------ | ----------------------- | ------------- |
-| fillStyle                | Attribute getter/setter | Implemented   |
-| strokeStyle              | Attribute getter/setter | Implemented   |
-| shadowColor              | Attribute getter/setter | **Undefined** |
-| shadowBlur               | Attribute getter/setter | **Undefined** |
-| shadowOffsetX            | Attribute getter/setter | **Undefined** |
-| shadowOffsetY            | Attribute getter/setter | **Undefined** |
-| createLinearGradient()   | Method                  | Implemented   |
-| createPattern()          | Method                  | Implemented   |
-| createRadialGradient()   | Method                  | Implemented   |
-| addColorStop()           | Method                  | Implemented   |
-| isPointInPath()          | Method                  | *Empty*       |
-| createEvent()            | Method                  | **Undefined** |
-| toDataURL()              | Method                  | **Undefined** |
-| lineCap                  | Attribute getter/setter | Implemented   |
-| lineJoin                 | Attribute getter/setter | Implemented   |
-| lineWidth                | Attribute getter/setter | Implemented   |
-| miterLimit               | Attribute getter/setter | Implemented   |
-| font                     | Attribute getter/setter | Implemented   |
-| textAlign                | Attribute getter/setter | Implemented   |
-| textBaseline             | Attribute getter/setter | Implemented   |
-| globalAlpha              | Attribute getter/setter | Implemented   |
-| globalCompositeOperation | Attribute getter/setter | Implemented   |
-| rect()                   | Method                  | Implemented   |
-| fillRect()               | Method                  | Implemented   |
-| strokeRect()             | Method                  | Implemented   |
-| clearRect()              | Method                  | Implemented   |
-| fill()                   | Method                  | Implemented   |
-| stroke()                 | Method                  | Implemented   |
-| beginPath()              | Method                  | Implemented   |
-| moveTo()                 | Method                  | Implemented   |
-| closePath()              | Method                  | Implemented   |
-| lineTo()                 | Method                  | Implemented   |
-| clip()                   | Method                  | Implemented   |
-| quadraticCurveTo()       | Method                  | Implemented   |
-| bezierCurveTo()          | Method                  | Implemented   |
-| arc()                    | Method                  | Implemented   |
-| arcTo()                  | Method                  | Implemented   |
-| scale()                  | Method                  | Implemented   |
-| rotate()                 | Method                  | Implemented   |
-| translate()              | Method                  | Implemented   |
-| transform()              | Method                  | Implemented   |
-| setTransform()           | Method                  | Implemented   |
-| fillText()               | Method                  | Implemented   |
-| strokeText()             | Method                  | Implemented   |
-| measureText()            | Method                  | *Empty*       |
-| drawImage()              | Method                  | Implemented   |
-| createImageData()        | Method                  | Implemented   |
-| getImageData()           | Method                  | Implemented   |
-| putImageData()           | Method                  | Implemented   |
-| save()                   | Method                  | Implemented   |
-| restore()                | Method                  | Implemented   |
-| getContext()             | Method                  | Implemented   |
-| loadTexture()            | Method                  | Implemented   |
-| unloadTexture()          | Method                  | Implemented   |
-| resetTransform()         | Method                  | Implemented   |
-| render()                 | Method                  | Implemented   |
-| capture()                | Method                  | *Empty*       |
-| resetClip()              | Method                  | Implemented   |
-
-
-### 2.ImageData APIs
-
-| API Name | API Type                | Status      |
-| -------- | ----------------------- | ----------- |
-| width    | Attribute getter/setter | Implemented |
-| height   | Attribute getter/setter | Implemented |
-| data     | Attribute getter/setter | Undefined   |
-| src      | Attribute getter/setter | Implemented |
-| complete | Attribute getter/setter | Implemented |
