--- GCanvas/docs/playground.md
@@ -1,26 +1,0 @@
-# Playground
-
-Download the playground app and visit our [playground](https://jsplayground.taobao.org/vueplayground) site.
-
-<iframe class="playground" src="https://jsplayground.taobao.org/vueplayground/"></iframe>
-
-# Android
-
-#### For Weex
-
-Click link below to download apk.
-
-[Download Weex Playground Debug Apk](https://raw.githubusercontent.com/g-platform/gcanvas-playground-android/master/weex_playground-debug.apk)
-
-# iOS
-
-#### For Weex
-
-Click link below to download ipa.
-
-<a href="itms-services://?action=download-manifest&url=https://raw.githubusercontent.com/g-platform/gcanvas-playground-ios/master/GCanvasPlayground-Releasedownload.plist" >Download Weex Playground  IPA</a>
-
-
-
-
-
