+++ tools/build_website/README.md
@@ -1,0 +1,4 @@
+```
+npm install
+npm run build
+```
