+++ tools/build_website/index.yaml
@@ -1,0 +1,17 @@
+
+docs:
+
+  Introduction: Brief.md
+  Guide:
+    Developer's Guide: Guide_Developer.md
+    Integrate GCanvas on Weex: Guide_Setup_Weex.md
+    Integrate GCanvas on ReactNative: Guide_Setup_ReactNative.md
+    Custom Bridge: Guide_Custom_Bridge.md
+    Contribute to GCanvas: Contributing.md
+  Design:
+    Architecture: An_Introduction_to_GCanvas.md
+  Reference:
+    Graphics 2D: Reference_Graphics_2D.md
+    WebGL: Reference_WebGL.md
+  About: About.md
+
