+++ tools/build_website/template/footer.pug
@@ -3,4 +3,4 @@
         a.open-source-icon(href="http://opensource.alibaba.com/")
             img(src="http://zos.alipayobjects.com/rmsportal/OCGqTPRSXYMsKjGiHPuP.png")
         div.license Released under Apache License
-        div.copyright Copyright ©️ 2018 Alibaba Inc.
+        div.copyright Copyright ©️ 2019 Alibaba Inc.
