--- website-src/index.yaml
@@ -1,17 +1,0 @@
-
-docs:
-
-  Introduction: ../GCanvas/docs/Brief.md
-  Guide:
-    Developer's Guide: ../GCanvas/docs/Guide_Developer.md
-    Integrate GCanvas on Weex: ../GCanvas/docs/Guide_Setup_Weex.md
-    Integrate GCanvas on ReactNative: ../GCanvas/docs/Guide_Setup_ReactNative.md
-    Custom Bridge: ../GCanvas/docs/Guide_Custom_Bridge.md
-    Contribute to GCanvas: ../GCanvas/docs/Contributing.md
-  Design:
-    Architecture: ../GCanvas/docs/An_Introduction_to_GCanvas.md
-  Reference:
-    Graphics 2D: ../GCanvas/docs/Reference_Graphics_2D.md
-    WebGL: ../GCanvas/docs/Reference_WebGL.md
-  About: ../GCanvas/docs/About.md
-
